refactor tests for "stateful" feature
