+++ components/nimbus/src/tests/test_enrollment.rs
@@ -5,122 +5,13 @@
 // Testing enrollment.rs
 
 use crate::{
-    defaults::Defaults, enrollment::*, error::Result, AppContext, AvailableRandomizationUnits,
-    Branch, BucketConfig, Experiment, FeatureConfig, NimbusTargetingHelper, TargetingAttributes,
+    defaults::Defaults, enrollment::*, error::Result, tests::helpers::get_test_experiments,
+    AppContext, AvailableRandomizationUnits, Branch, BucketConfig, Experiment, FeatureConfig,
+    NimbusTargetingHelper, TargetingAttributes,
 };
 use serde_json::{json, Value};
 use std::collections::{HashMap, HashSet};
 use uuid::Uuid;
-
-cfg_if::cfg_if! {
-    if #[cfg(feature = "stateful")] {
-        use crate::behavior::EventStore;
-        use std::sync::{Arc, Mutex};
-    }
-}
-
-fn get_test_experiments() -> Vec<Experiment> {
-    vec![
-        serde_json::from_value(json!({
-            "schemaVersion": "1.0.0",
-            "slug": "secure-gold",
-            "endDate": null,
-            "featureIds": ["some_control"],
-            "branches":[
-                {
-                    "slug": "control",
-                    "ratio": 1,
-                    "feature": {
-                        "featureId": "some_control",
-                        "enabled": false,
-                        "value": {
-                            "text": "OK then",
-                            "number": 42
-                        }
-                    }
-                },
-                {
-                    "slug": "treatment",
-                    "ratio":1,
-                    "feature": {
-                        "featureId": "some_control",
-                        "enabled": true,
-                        "value": {
-                            "text": "OK then",
-                            "number": 42
-                        }
-                    }
-                }
-            ],
-            "channel": "nightly",
-            "probeSets":[],
-            "startDate":null,
-            "appName": "fenix",
-            "appId": "org.mozilla.fenix",
-            "bucketConfig":{
-                // Setup to enroll everyone by default.
-                "count":10_000,
-                "start":0,
-                "total":10_000,
-                "namespace":"secure-gold",
-                "randomizationUnit":"nimbus_id"
-            },
-            "userFacingName":"Diagnostic test experiment",
-            "referenceBranch":"control",
-            "isEnrollmentPaused":false,
-            "proposedEnrollment":7,
-            "userFacingDescription":"This is a test experiment for diagnostic purposes.",
-            "id":"secure-gold",
-            "last_modified":1_602_197_324_372i64
-        }))
-        .unwrap(),
-        serde_json::from_value(json!({
-            "schemaVersion": "1.0.0",
-            "slug": "secure-silver",
-            "endDate": null,
-            "branches":[
-                {
-                    "slug": "control",
-                    "ratio": 1,
-                    "feature": {
-                        "featureId": "about_welcome",
-                        "enabled": true,
-                    }
-                },
-                {
-                    "slug": "treatment",
-                    "ratio": 1,
-                    "feature": {
-                        "featureId": "about_welcome",
-                        "enabled": false,
-                    }
-                },
-            ],
-            "featureIds": ["about_welcome"],
-            "channel": "nightly",
-            "probeSets":[],
-            "startDate":null,
-            "appName":"fenix",
-            "appId":"org.mozilla.fenix",
-            "bucketConfig":{
-                // Also enroll everyone.
-                "count":10_000,
-                "start":0,
-                "total":10_000,
-                "namespace":"secure-silver",
-                "randomizationUnit":"nimbus_id"
-            },
-            "userFacingName":"2nd test experiment",
-            "referenceBranch":"control",
-            "isEnrollmentPaused":false,
-            "proposedEnrollment":7,
-            "userFacingDescription":"2nd test experiment.",
-            "id":"secure-silver",
-            "last_modified":1_602_197_324_372i64
-        }))
-        .unwrap(),
-    ]
-}
 
 fn get_feature_conflict_test_experiments() -> Vec<Experiment> {
     vec![
@@ -498,34 +389,6 @@ fn enrollment_evolver<'a>(
     aru: &'a AvailableRandomizationUnits,
 ) -> EnrollmentsEvolver<'a> {
     EnrollmentsEvolver::new(nimbus_id, aru, targeting_helper)
-}
-
-impl From<TargetingAttributes> for NimbusTargetingHelper {
-    fn from(value: TargetingAttributes) -> Self {
-        cfg_if::cfg_if! {
-            if #[cfg(feature = "stateful")] {
-                let store = Arc::new(Mutex::new(EventStore::new()));
-                NimbusTargetingHelper::new(value, store)
-            } else {
-                NimbusTargetingHelper::new(value)
-            }
-        }
-    }
-}
-
-impl Default for NimbusTargetingHelper {
-    fn default() -> Self {
-        cfg_if::cfg_if! {
-            if #[cfg(feature = "stateful")] {
-                let ctx: AppContext = Default::default();
-                let store = Arc::new(Mutex::new(EventStore::new()));
-                NimbusTargetingHelper::new(ctx, store)
-            } else {
-                let ctx: AppContext = Default::default();
-                NimbusTargetingHelper::new(ctx)
-            }
-        }
-    }
 }
 
 #[test]
@@ -948,7 +811,7 @@ fn test_evolver_experiment_update_enrolled_then_branches_changed() -> Result<()>
 #[test]
 fn test_evolver_experiment_update_enrolled_then_branch_disappears() -> Result<()> {
     let mut exp = get_test_experiments()[0].clone();
-    exp.branches = vec![crate::Branch {
+    exp.branches = vec![Branch {
         slug: "bobo-branch".to_owned(),
         ratio: 1,
         feature: None,
@@ -1097,7 +960,7 @@ fn test_evolver_feature_can_have_only_one_experiment() -> Result<()> {
 
     let enrolled_feature = features.get("about_welcome").unwrap();
     assert_eq!(
-        serde_json::Value::Object(enrolled_feature.feature.value.clone()),
+        Value::Object(enrolled_feature.feature.value.clone()),
         json!({ "text": "OK then", "number": 42})
     );
 
@@ -2500,377 +2363,4 @@ fn test_filter_experiments_by_closure() -> Result<()> {
     assert_eq!(exp_enrollments[0].slug, experiment.slug);
 
     Ok(())
-}
-
-// Older tests that also use the DB.
-// XXX: make them less complicated (since the transitions are covered above), just see if we write to the DB properly.
-#[cfg(feature = "stateful")]
-mod stateful_tests {
-    use super::*;
-    use crate::persistence::{Database, Readable, StoreId};
-
-    fn get_experiment_enrollments<'r>(
-        db: &Database,
-        reader: &'r impl Readable<'r>,
-    ) -> Result<Vec<ExperimentEnrollment>> {
-        db.get_store(StoreId::Enrollments).collect_all(reader)
-    }
-
-    impl From<EventStore> for NimbusTargetingHelper {
-        fn from(value: EventStore) -> Self {
-            let ctx: AppContext = Default::default();
-            NimbusTargetingHelper::new(ctx, Arc::new(Mutex::new(value)))
-        }
-    }
-
-    #[test]
-    fn test_enrollments() -> Result<()> {
-        let _ = env_logger::try_init();
-        let tmp_dir = tempfile::tempdir()?;
-        let db = Database::new(&tmp_dir)?;
-        let mut writer = db.write()?;
-        let exp1 = get_test_experiments()[0].clone();
-        let nimbus_id = Uuid::new_v4();
-        let aru = Default::default();
-        let targeting_attributes = AppContext {
-            app_name: "fenix".to_string(),
-            app_id: "org.mozilla.fenix".to_string(),
-            channel: "nightly".to_string(),
-            ..Default::default()
-        }
-        .into();
-        assert_eq!(get_enrollments(&db, &writer)?.len(), 0);
-
-        let evolver = EnrollmentsEvolver::new(&nimbus_id, &aru, &targeting_attributes);
-        let events = evolver.evolve_enrollments_in_db(&db, &mut writer, &[exp1])?;
-
-        let enrollments = get_enrollments(&db, &writer)?;
-        assert_eq!(enrollments.len(), 1);
-        let enrollment = &enrollments[0];
-        assert_eq!(enrollment.slug, "secure-gold");
-        assert_eq!(enrollment.user_facing_name, "Diagnostic test experiment");
-        assert_eq!(
-            enrollment.user_facing_description,
-            "This is a test experiment for diagnostic purposes."
-        );
-        assert!(enrollment.branch_slug == "control" || enrollment.branch_slug == "treatment");
-        // Ensure the event was created too.
-        assert_eq!(events.len(), 1);
-        let event = &events[0];
-        assert_eq!(event.experiment_slug, "secure-gold");
-        assert!(event.branch_slug == "control" || event.branch_slug == "treatment");
-        assert!(matches!(
-            event.change,
-            EnrollmentChangeEventType::Enrollment
-        ));
-
-        // Get the ExperimentEnrollment from the DB.
-        let ee: ExperimentEnrollment = db
-            .get_store(StoreId::Enrollments)
-            .get(&writer, "secure-gold")?
-            .expect("should exist");
-        assert!(matches!(
-            ee.status,
-            EnrollmentStatus::Enrolled {
-                reason: EnrolledReason::Qualified,
-                ..
-            }
-        ));
-
-        // Now opt-out.
-        opt_out(&db, &mut writer, "secure-gold")?;
-        assert_eq!(get_enrollments(&db, &writer)?.len(), 0);
-        // check we recorded the "why" correctly.
-        let ee: ExperimentEnrollment = db
-            .get_store(StoreId::Enrollments)
-            .get(&writer, "secure-gold")?
-            .expect("should exist");
-        assert!(matches!(
-            ee.status,
-            EnrollmentStatus::Disqualified {
-                reason: DisqualifiedReason::OptOut,
-                ..
-            }
-        ));
-
-        // Opt in to a specific branch.
-        opt_in_with_branch(&db, &mut writer, "secure-gold", "treatment")?;
-        let enrollments = get_enrollments(&db, &writer)?;
-        assert_eq!(enrollments.len(), 1);
-        let enrollment = &enrollments[0];
-        assert_eq!(enrollment.slug, "secure-gold");
-        assert!(enrollment.branch_slug == "treatment");
-
-        writer.commit()?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_updates() -> Result<()> {
-        let _ = env_logger::try_init();
-        let tmp_dir = tempfile::tempdir()?;
-        let db = Database::new(&tmp_dir)?;
-        let mut writer = db.write()?;
-        let nimbus_id = Uuid::new_v4();
-        let aru = Default::default();
-        let th = AppContext {
-            app_name: "fenix".to_string(),
-            app_id: "org.mozilla.fenix".to_string(),
-            channel: "nightly".to_string(),
-            ..Default::default()
-        }
-        .into();
-        assert_eq!(get_enrollments(&db, &writer)?.len(), 0);
-        let exps = get_test_experiments();
-
-        let evolver = EnrollmentsEvolver::new(&nimbus_id, &aru, &th);
-        let events = evolver.evolve_enrollments_in_db(&db, &mut writer, &exps)?;
-
-        let enrollments = get_enrollments(&db, &writer)?;
-        assert_eq!(enrollments.len(), 2);
-        assert_eq!(events.len(), 2);
-
-        // pretend we just updated from the server and one of the 2 is missing.
-        let exps = &[exps[1].clone()];
-        let evolver = EnrollmentsEvolver::new(&nimbus_id, &aru, &th);
-        let events = evolver.evolve_enrollments_in_db(&db, &mut writer, exps)?;
-
-        // should only have 1 now.
-        let enrollments = get_enrollments(&db, &writer)?;
-        assert_eq!(enrollments.len(), 1);
-        // Check that the un-enrolled event was emitted.
-        assert_eq!(events.len(), 1);
-        let event = &events[0];
-        assert_eq!(event.experiment_slug, "secure-gold");
-        assert!(matches!(
-            event.change,
-            EnrollmentChangeEventType::Unenrollment
-        ));
-
-        writer.commit()?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_global_opt_out() -> Result<()> {
-        let _ = env_logger::try_init();
-        let tmp_dir = tempfile::tempdir()?;
-        let db = Database::new(&tmp_dir)?;
-        let mut writer = db.write()?;
-        let nimbus_id = Uuid::new_v4();
-        let th = AppContext {
-            app_name: "fenix".to_string(),
-            app_id: "org.mozilla.fenix".to_string(),
-            channel: "nightly".to_string(),
-            ..Default::default()
-        }
-        .into();
-        let aru = Default::default();
-        assert_eq!(get_enrollments(&db, &writer)?.len(), 0);
-        let exps = get_test_experiments();
-
-        // User has opted out of new experiments.
-        set_global_user_participation(&db, &mut writer, false)?;
-
-        let evolver = EnrollmentsEvolver::new(&nimbus_id, &aru, &th);
-        let events = evolver.evolve_enrollments_in_db(&db, &mut writer, &exps)?;
-
-        let enrollments = get_enrollments(&db, &writer)?;
-        assert_eq!(enrollments.len(), 0);
-        assert!(events.is_empty());
-        // We should see the experiment non-enrollments.
-        assert_eq!(get_experiment_enrollments(&db, &writer)?.len(), 2);
-        let num_not_enrolled_enrollments = get_experiment_enrollments(&db, &writer)?
-            .into_iter()
-            .filter(|enr| {
-                matches!(
-                    enr.status,
-                    EnrollmentStatus::NotEnrolled {
-                        reason: NotEnrolledReason::OptOut
-                    }
-                )
-            })
-            .count();
-        assert_eq!(num_not_enrolled_enrollments, 2);
-
-        // User opts in, and updating should enroll us in 2 experiments.
-        set_global_user_participation(&db, &mut writer, true)?;
-
-        let evolver = EnrollmentsEvolver::new(&nimbus_id, &aru, &th);
-        let events = evolver.evolve_enrollments_in_db(&db, &mut writer, &exps)?;
-
-        let enrollments = get_enrollments(&db, &writer)?;
-        assert_eq!(enrollments.len(), 2);
-        assert_eq!(events.len(), 2);
-        // We should see 2 experiment enrollments.
-        assert_eq!(get_experiment_enrollments(&db, &writer)?.len(), 2);
-        let num_enrolled_enrollments = get_experiment_enrollments(&db, &writer)?
-            .into_iter()
-            .filter(|enr| matches!(enr.status, EnrollmentStatus::Enrolled { .. }))
-            .count();
-        assert_eq!(num_enrolled_enrollments, 2);
-
-        // Opting out and updating should give us two disqualified enrollments
-        set_global_user_participation(&db, &mut writer, false)?;
-
-        let evolver = EnrollmentsEvolver::new(&nimbus_id, &aru, &th);
-        let events = evolver.evolve_enrollments_in_db(&db, &mut writer, &exps)?;
-
-        let enrollments = get_enrollments(&db, &writer)?;
-        assert_eq!(enrollments.len(), 0);
-        assert_eq!(events.len(), 2);
-        // We should see 2 experiment enrolments, this time they're both opt outs
-        assert_eq!(get_experiment_enrollments(&db, &writer)?.len(), 2);
-
-        assert_eq!(
-            get_experiment_enrollments(&db, &writer)?
-                .into_iter()
-                .filter(|enr| {
-                    matches!(
-                        enr.status,
-                        EnrollmentStatus::Disqualified {
-                            reason: DisqualifiedReason::OptOut,
-                            ..
-                        }
-                    )
-                })
-                .count(),
-            2
-        );
-
-        // Opting in again and updating SHOULD NOT enroll us again (we've been disqualified).
-        set_global_user_participation(&db, &mut writer, true)?;
-
-        let evolver = EnrollmentsEvolver::new(&nimbus_id, &aru, &th);
-        let events = evolver.evolve_enrollments_in_db(&db, &mut writer, &exps)?;
-
-        let enrollments = get_enrollments(&db, &writer)?;
-        assert_eq!(enrollments.len(), 0);
-        assert!(events.is_empty());
-
-        assert_eq!(
-            get_experiment_enrollments(&db, &writer)?
-                .into_iter()
-                .filter(|enr| {
-                    matches!(
-                        enr.status,
-                        EnrollmentStatus::Disqualified {
-                            reason: DisqualifiedReason::OptOut,
-                            ..
-                        }
-                    )
-                })
-                .count(),
-            2
-        );
-
-        writer.commit()?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_telemetry_reset() -> Result<()> {
-        let _ = env_logger::try_init();
-        let tmp_dir = tempfile::tempdir()?;
-        let db = Database::new(&tmp_dir)?;
-        let mut writer = db.write()?;
-
-        let mock_exp1_slug = "exp-1".to_string();
-        let mock_exp1_branch = "branch-1".to_string();
-        let mock_exp2_slug = "exp-2".to_string();
-        let mock_exp2_branch = "branch-2".to_string();
-        let mock_exp3_slug = "exp-3".to_string();
-
-        // Three currently-known experiments, in different states.
-        let store = db.get_store(StoreId::Enrollments);
-        store.put(
-            &mut writer,
-            &mock_exp1_slug,
-            &ExperimentEnrollment {
-                slug: mock_exp1_slug.clone(),
-                status: EnrollmentStatus::new_enrolled(
-                    EnrolledReason::Qualified,
-                    &mock_exp1_branch,
-                ),
-            },
-        )?;
-        store.put(
-            &mut writer,
-            &mock_exp2_slug,
-            &ExperimentEnrollment {
-                slug: mock_exp2_slug.clone(),
-                status: EnrollmentStatus::Disqualified {
-                    reason: DisqualifiedReason::Error,
-                    branch: mock_exp2_branch.clone(),
-                    enrollment_id: Uuid::new_v4(),
-                },
-            },
-        )?;
-        store.put(
-            &mut writer,
-            &mock_exp3_slug,
-            &ExperimentEnrollment {
-                slug: mock_exp3_slug.clone(),
-                status: EnrollmentStatus::NotEnrolled {
-                    reason: NotEnrolledReason::NotTargeted,
-                },
-            },
-        )?;
-        writer.commit()?;
-
-        let mut writer = db.write()?;
-        let events = reset_telemetry_identifiers(&db, &mut writer)?;
-        writer.commit()?;
-
-        let enrollments = db.collect_all::<ExperimentEnrollment>(StoreId::Enrollments)?;
-        assert_eq!(enrollments.len(), 3);
-
-        // The enrolled experiment should have moved to disqualified with nil enrollment_id.
-        assert_eq!(enrollments[0].slug, mock_exp1_slug);
-        assert!(
-            matches!(&enrollments[0].status, EnrollmentStatus::Disqualified {
-            reason: DisqualifiedReason::OptOut,
-            branch,
-            enrollment_id,
-            ..
-        } if *branch == mock_exp1_branch && enrollment_id.is_nil())
-        );
-
-        // The disqualified experiment should have stayed disqualified, with nil enrollment_id.
-        assert_eq!(enrollments[1].slug, mock_exp2_slug);
-        assert!(
-            matches!(&enrollments[1].status, EnrollmentStatus::Disqualified {
-            reason: DisqualifiedReason::Error,
-            branch,
-            enrollment_id,
-            ..
-        } if *branch == mock_exp2_branch && enrollment_id.is_nil())
-        );
-
-        // The not-enrolled experiment should have been unchanged.
-        assert_eq!(enrollments[2].slug, mock_exp3_slug);
-        assert!(matches!(
-            &enrollments[2].status,
-            EnrollmentStatus::NotEnrolled {
-                reason: NotEnrolledReason::NotTargeted,
-                ..
-            }
-        ));
-
-        // We should have returned a single disqualification event.
-        assert_eq!(events.len(), 1);
-        assert!(matches!(&events[0], EnrollmentChangeEvent {
-            change: EnrollmentChangeEventType::Disqualification,
-            reason: Some(reason),
-            experiment_slug,
-            branch_slug,
-            enrollment_id,
-        } if reason == "optout"
-            && *experiment_slug == mock_exp1_slug
-            && *branch_slug == mock_exp1_branch
-            && ! Uuid::parse_str(enrollment_id)?.is_nil()
-        ));
-
-        Ok(())
-    }
 }
